Moves some packets
