+++ Projects/Server/Network/PacketHandlers.cs
@@ -49,7 +49,7 @@ namespace Server.Network
     Encoded = 0xC0
   }
 
-  public static class PacketHandlers
+  public static partial class PacketHandlers
   {
     private const int m_AuthIDWindowSize = 128;
     private static readonly PacketHandler[] m_6017Handlers = new PacketHandler[0x100];
@@ -840,10 +840,10 @@ namespace Server.Network
       }
     }
 
-    public static void Disconnect(NetState state, BufferReader reader)
-    {
-      var minusOne = reader.ReadInt32();
-    }
+    // public static void Disconnect(NetState state, BufferReader reader)
+    // {
+    //   var minusOne = reader.ReadInt32();
+    // }
 
     public static void LiftReq(NetState state, BufferReader reader)
     {
@@ -1987,250 +1987,6 @@ namespace Server.Network
       EventSink.InvokeLogin(m);
 
       m.ClearFastwalkStack();
-    }
-
-    public static void CreateCharacter(NetState state, BufferReader reader)
-    {
-      var unk1 = reader.ReadInt32();
-      var unk2 = reader.ReadInt32();
-      int unk3 = reader.ReadByte();
-      var name = reader.ReadAscii(30);
-
-      reader.Seek(2, SeekOrigin.Current);
-      var flags = reader.ReadInt32();
-      reader.Seek(8, SeekOrigin.Current);
-      int prof = reader.ReadByte();
-      reader.Seek(15, SeekOrigin.Current);
-
-      int genderRace = reader.ReadByte();
-
-      int str = reader.ReadByte();
-      int dex = reader.ReadByte();
-      int intl = reader.ReadByte();
-      int is1 = reader.ReadByte();
-      int vs1 = reader.ReadByte();
-      int is2 = reader.ReadByte();
-      int vs2 = reader.ReadByte();
-      int is3 = reader.ReadByte();
-      int vs3 = reader.ReadByte();
-      int hue = reader.ReadUInt16();
-      int hairVal = reader.ReadInt16();
-      int hairHue = reader.ReadInt16();
-      int hairValf = reader.ReadInt16();
-      int hairHuef = reader.ReadInt16();
-      reader.ReadByte();
-      int cityIndex = reader.ReadByte();
-      var charSlot = reader.ReadInt32();
-      var clientIP = reader.ReadInt32();
-      int shirtHue = reader.ReadInt16();
-      int pantsHue = reader.ReadInt16();
-
-      /*
-      Pre-7.0.0.0:
-      0x00, 0x01 -> Human Male, Human Female
-      0x02, 0x03 -> Elf Male, Elf Female
-
-      Post-7.0.0.0:
-      0x00, 0x01
-      0x02, 0x03 -> Human Male, Human Female
-      0x04, 0x05 -> Elf Male, Elf Female
-      0x05, 0x06 -> Gargoyle Male, Gargoyle Female
-      */
-
-      var female = genderRace % 2 != 0;
-
-      Race race;
-
-      if (state.StygianAbyss)
-      {
-        var raceID = (byte)(genderRace < 4 ? 0 : genderRace / 2 - 1);
-        race = Race.Races[raceID];
-      }
-      else
-      {
-        race = Race.Races[(byte)(genderRace / 2)];
-      }
-
-      race ??= Race.DefaultRace;
-
-      var info = state.CityInfo;
-      var a = state.Account;
-
-      if (info == null || a == null || cityIndex < 0 || cityIndex >= info.Length)
-      {
-        state.Dispose();
-      }
-      else
-      {
-        // Check if anyone is using this account
-        for (var i = 0; i < a.Length; ++i)
-        {
-          var check = a[i];
-
-          if (check != null && check.Map != Map.Internal)
-          {
-            Console.WriteLine("Login: {0}: Account in use", state);
-            state.Send(new PopupMessage(PMMessage.CharInWorld));
-            return;
-          }
-        }
-
-        state.Flags = (ClientFlags)flags;
-
-        var args = new CharacterCreatedEventArgs(
-          state, a,
-          name, female, hue,
-          str, dex, intl,
-          info[cityIndex],
-          new[]
-          {
-            new SkillNameValue((SkillName)is1, vs1),
-            new SkillNameValue((SkillName)is2, vs2),
-            new SkillNameValue((SkillName)is3, vs3)
-          },
-          shirtHue, pantsHue,
-          hairVal, hairHue,
-          hairValf, hairHuef,
-          prof,
-          race);
-
-        state.Send(new ClientVersionReq());
-
-        state.BlockAllPackets = true;
-
-        EventSink.InvokeCharacterCreated(args);
-
-        var m = args.Mobile;
-
-        if (m != null)
-        {
-          state.Mobile = m;
-          m.NetState = state;
-          new LoginTimer(state, m).Start();
-        }
-        else
-        {
-          state.BlockAllPackets = false;
-          state.Dispose();
-        }
-      }
-    }
-
-    public static void CreateCharacter70160(NetState state, BufferReader reader)
-    {
-      var unk1 = reader.ReadInt32();
-      var unk2 = reader.ReadInt32();
-      int unk3 = reader.ReadByte();
-      var name = reader.ReadAscii(30);
-
-      reader.Seek(2, SeekOrigin.Current);
-      var flags = reader.ReadInt32();
-      reader.Seek(8, SeekOrigin.Current);
-      int prof = reader.ReadByte();
-      reader.Seek(15, SeekOrigin.Current);
-
-      int genderRace = reader.ReadByte();
-
-      int str = reader.ReadByte();
-      int dex = reader.ReadByte();
-      int intl = reader.ReadByte();
-      int is1 = reader.ReadByte();
-      int vs1 = reader.ReadByte();
-      int is2 = reader.ReadByte();
-      int vs2 = reader.ReadByte();
-      int is3 = reader.ReadByte();
-      int vs3 = reader.ReadByte();
-      int is4 = reader.ReadByte();
-      int vs4 = reader.ReadByte();
-
-      int hue = reader.ReadUInt16();
-      int hairVal = reader.ReadInt16();
-      int hairHue = reader.ReadInt16();
-      int hairValf = reader.ReadInt16();
-      int hairHuef = reader.ReadInt16();
-      reader.ReadByte();
-      int cityIndex = reader.ReadByte();
-      var charSlot = reader.ReadInt32();
-      var clientIP = reader.ReadInt32();
-      int shirtHue = reader.ReadInt16();
-      int pantsHue = reader.ReadInt16();
-
-      /*
-      0x00, 0x01
-      0x02, 0x03 -> Human Male, Human Female
-      0x04, 0x05 -> Elf Male, Elf Female
-      0x05, 0x06 -> Gargoyle Male, Gargoyle Female
-      */
-
-      var female = genderRace % 2 != 0;
-
-      Race race;
-
-      var raceID = (byte)(genderRace < 4 ? 0 : genderRace / 2 - 1);
-      race = Race.Races[raceID] ?? Race.DefaultRace;
-
-      var info = state.CityInfo;
-      var a = state.Account;
-
-      if (info == null || a == null || cityIndex < 0 || cityIndex >= info.Length)
-      {
-        state.Dispose();
-      }
-      else
-      {
-        // Check if anyone is using this account
-        for (var i = 0; i < a.Length; ++i)
-        {
-          var check = a[i];
-
-          if (check != null && check.Map != Map.Internal)
-          {
-            Console.WriteLine("Login: {0}: Account in use", state);
-            state.Send(new PopupMessage(PMMessage.CharInWorld));
-            return;
-          }
-        }
-
-        state.Flags = (ClientFlags)flags;
-
-        var args = new CharacterCreatedEventArgs(
-          state, a,
-          name, female, hue,
-          str, dex, intl,
-          info[cityIndex],
-          new[]
-          {
-            new SkillNameValue((SkillName)is1, vs1),
-            new SkillNameValue((SkillName)is2, vs2),
-            new SkillNameValue((SkillName)is3, vs3),
-            new SkillNameValue((SkillName)is4, vs4)
-          },
-          shirtHue, pantsHue,
-          hairVal, hairHue,
-          hairValf, hairHuef,
-          prof,
-          race);
-
-        state.Send(new ClientVersionReq());
-
-        state.BlockAllPackets = true;
-
-        EventSink.InvokeCharacterCreated(args);
-
-        var m = args.Mobile;
-
-        if (m != null)
-        {
-          state.Mobile = m;
-          m.NetState = state;
-          new LoginTimer(state, m).Start();
-        }
-        else
-        {
-          state.BlockAllPackets = false;
-          state.Dispose();
-        }
-      }
     }
 
     private static int GenerateAuthID(NetState state)
